Conflicts:
	Engine/AppInstance.h
	Engine/AppManager.cpp
	Engine/AppManager.h
	Engine/Knob.cpp
	Engine/Knob.h
	Engine/KnobSerialization.cpp
	Engine/Node.cpp
	Engine/OfxParamInstance.cpp
	Engine/OutputSchedulerThread.cpp
	Engine/RotoContextPrivate.h
	Engine/ViewerInstance.cpp
	Engine/ViewerInstance.h
	Engine/ViewerInstancePrivate.h
	Gui/DockablePanel.cpp
	Gui/Gui.cpp
	Gui/GuiAppInstance.cpp
	Gui/GuiAppInstance.h
	Gui/InfoViewerWidget.h
	Gui/KnobGui.cpp
	Gui/KnobGuiTypes.cpp
	Gui/MultiInstancePanel.cpp
	Gui/NodeGui.cpp
	Gui/ProjectGuiSerialization.cpp
	Gui/RotoPanel.cpp
	Gui/TabWidget.cpp
	Gui/ViewerTab.cpp
	Gui/ViewerTab.h

+++ Gui/CustomParamInteract.cpp
@@ -63,7 +63,7 @@ CustomParamInteract::CustomParamInteract(KnobGui* knob,
     : QGLWidget(parent)
       , _imp( new CustomParamInteractPrivate(knob,ofxParamHandle,entryPoint) )
 {
-    int minW,minH;
+    double minW,minH;
 
     entryPoint->getMinimumSize(minW, minH);
     setMinimumSize(minW, minH);
